added freecam

+++ src/main/kotlin/com/buttersus/wiremaster/extensions/BooleanExtensions.kt
@@ -1,5 +1,5 @@
 package com.buttersus.wiremaster.extensions
 
 @Suppress("unused")
-fun Boolean.toInt() =
-    if (this) 1 else 0
+fun Boolean.toDouble() =
+    if (this) 1.0 else 0.0
